agents/tester/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "meta-agent-tester"
version = "0.1.0"
description = "Test generator agent"
requires-python = ">=3.12"
dependencies = [
    "google-adk[a2a]>=0.5.0",
    "meta-agent-shared",
]
